buildingblocks/requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
black==21.7b0
boto3==1.18.0
botocore==1.21.0
click==8.0.1
flake8==3.9.2
flask==2.0.1
flask-dynamo==0.1.2
flask_restful==0.3.9
flask_cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.2
pytz==2021.1
regex==2021.7.6
s3transfer==0.5.0
six==1.16.0
tomli==1.1.0
urllib3==1.26.6
Werkzeug==2.0.1
pandas==1.3.1
SQLAlchemy==1.4.22
python-jose==3.3.0
flask_lambda==0.0.4
kanpai==0.1.15
pdfkit==0.6.1
headless-pdfkit==0.1.1
xhtml2pdf==0.2.5